update jump.h fix Jget_flt typedef

+++ jump.h
@@ -9,7 +9,7 @@ typedef int (*fn3)(int , int , int);
 typedef int (*fn4)(char*);
 typedef int (*fn5)(int , int , int, int);
 typedef char* (*fn6)(int);
-typedef int (*fn7)(double);
+typedef double (*fn7)(int);
 typedef void (*tpred)(char*, int(*pred)(int , int));
 typedef void (*tuser)(char*, int(*user)(int , int), int weight, int spec);
